requirements-core.txt
# NumPy 2.x wheels need x86-64-v2; pin 1.x for older VPS / emulated CPUs
numpy>=1.26.4,<2

# PDF Processing
pymupdf==1.25.3  # Fast PDF text extraction
pytesseract==0.3.13  # OCR when rulebook pages are image-only (needs tesseract binary)
Pillow>=10.4.0

# Vector Database & Embeddings
chromadb==0.5.21  # Local vector database (stores embeddings free/locally)
posthog==4.8.0  # Pin: Chroma telemetry breaks with newer posthog (#4997)
sentence-transformers==3.3.1  # Runs open-source embedding models locally

# AI Integration
google-generativeai==0.8.3  # Gemini API for the free tier generative completion

# Configuration & Utilities
python-dotenv==1.0.1  # For loading GEMINI_API_KEY securely

# UI / Web (Flask only in Docker; see requirements.txt for Streamlit)
flask==3.0.3
markdown>=3.5
bleach>=6.1.0
